pyproject.toml
[tool.poetry]
name = "psyai-chat"
version = "0.1.0"
description = "Django-based AI chat for psychological support (MVP)"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
channels = "^4.0.0"
daphne = "^4.1.2"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
python-dotenv = "^1.0.1"
openai = "^1.88.0"
pillow = "^11.2.1"
whitenoise = "^6.9.0"
celery = "^5.4.0"
redis = "^5.0.7"
yookassa = "^3.5.0"
django-anymail = "^13.0"
sendgrid = "^6.12.4"
boto3 = "^1.38.42"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-django = "^4.8.0"
ruff = "^0.4.4"
black = "^24.4.2"
pre-commit = "^3.7.0"
pytest-asyncio = "^0.23.7"
pytest-env = "^1.1.3"
flower = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.main:api"

# --- Инструменты форматирования и линтинга ---

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
]

# Исключаем папки, которые не нужно проверять
exclude = [
    "migrations",
]